requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-taggit==3.1.0
djangorestframework==3.14.0
fontawesomefree==6.2.1
install==1.3.5
pytz==2022.7.1
sqlparse==0.4.3
